+++ spring-batch-core/src/main/java/org/springframework/batch/core/step/skip/NeverSkipItemSkipPolicy.java
@@ -25,7 +25,7 @@ package org.springframework.batch.core.step.skip;
 public class NeverSkipItemSkipPolicy implements SkipPolicy{
 
 	@Override
-	public boolean shouldSkip(Throwable t, int skipCount) {
+	public boolean shouldSkip(Throwable t, long skipCount) {
 		return false;
 	}
